:construction: add methods to parse page, scrape shows, and prompt user
created a method that goes over all the shows on the website and puts them into an array, which is them displayed to the user to choose from. The prompt can be filtered by typing. Only problem right now is that the amount of stuff it looks at is insane.